backend/requirements.txt
asgiref==3.11.1
celery==5.4.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.7
daphne==4.2.1
colorama==0.4.6
coverage==7.13.5
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
iniconfig==2.3.0
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-django==4.12.0
reportlab==4.4.10
ruff==0.11.12
sentry-sdk==2.58.0
sqlparse==0.5.5
tzdata==2026.1
